Newest Anki asks for all front-cards to be different
Anki is trying to be smart and determine whether a card is repeated
(it doesn't really want two cards with different backs but the same
front). But this deck duplicates cards to drill down into the learner's
head the differences between minimal pair sounds. So, all cards have now
a hidden unique identifier (an HTML comment) to force Anki into
compliance!